Rename sbt-play-scalajs to sbt-web-scalajs because it only depends on sbt-web and Scala.js. It has no Play dependency. sbt-web-scalajs includes changes to facilitate the integration of Scala.js with any sbt-web server, Play, Akka HTTP, Spray etc.
Main change:
The user should now use the scalaJSPipeline task, which runs scalaJSDev in dev mode, runs scalaJSProd otherwise.

Other changes:
* `ScalaJSPlay` has been renamed to `ScalaJSWeb`
* `PlayScalaJS` has been renamed to `WebScalaJS`
* `scalaJSDev` type changed from `TaskKey[Seq[PathMapping]]` to `TaskKey[Pipeline.Stage]`
* `scalaJSTest` has been removed
