+++ README.md
@@ -1,53 +1,77 @@
-# sbt-play-scalajs
+# sbt-web-scalajs
 
 [![License](http://img.shields.io/:license-Apache%202-red.svg)](http://www.apache.org/licenses/LICENSE-2.0.txt)
-[![Download](https://api.bintray.com/packages/vmunier/scalajs/sbt-play-scalajs/images/download.svg) ](https://bintray.com/vmunier/scalajs/sbt-play-scalajs/_latestVersion)
-[![Gitter](https://badges.gitter.im/Join%20Chat.svg)](https://gitter.im/vmunier/sbt-play-scalajs?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge)
+[![Download](https://api.bintray.com/packages/vmunier/scalajs/sbt-web-scalajs/images/download.svg) ](https://bintray.com/vmunier/scalajs/sbt-web-scalajs/_latestVersion)
+[![Gitter](https://badges.gitter.im/Join%20Chat.svg)](https://gitter.im/vmunier/sbt-web-scalajs?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge)
 
-sbt-play-scalajs is a SBT plugin which allows you to use Scala.js along with Play Framework.
-It uses the [sbt-web](https://github.com/sbt/sbt-web) and [scala-js](https://github.com/scala-js/scala-js) plugins.
+sbt-web-scalajs is a SBT plugin which allows you to use Scala.js along with any sbt-web server. It uses the [sbt-web](https://github.com/sbt/sbt-web) and [scala-js](https://github.com/scala-js/scala-js) plugins.
 
 ## Setup
 
-Specify the sbt version in `project/build.properties` which needs to be 0.13.5 or higher:
+Specify the sbt version in `project/build.properties`, which needs to be 0.13.7 or higher:
 ```
-sbt.version=0.13.5
+sbt.version=0.13.7
 ```
 
-Add the sbt plugin to the `project/plugins.sbt` file along with Play! and Scala.js:
+Add the sbt plugin to the `project/plugins.sbt` file along with Scala.js:
 ```
-addSbtPlugin("com.vmunier" % "sbt-play-scalajs" % "0.3.1")
+addSbtPlugin("com.vmunier" % "sbt-web-scalajs" % "1.0.0")
 
-addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.5.4")
-
-addSbtPlugin("org.scala-js" % "sbt-scalajs" % "0.6.10")
+addSbtPlugin("org.scala-js" % "sbt-scalajs" % "0.6.12")
 ```
 
 Lastly, put the following configuration in `build.sbt`:
 ```
-import sbt.Project.projectToRef
+lazy val server = project.settings(
+  scalaJSProjects := Seq(client),
+  pipelineStages in Assets := Seq(scalaJSPipeline)
+).enablePlugins(SbtWeb)
 
-lazy val jsProjects = Seq(js)
-
-lazy val jvm = project.settings(
-  scalaJSProjects := jsProjects,
-  pipelineStages := Seq(scalaJSProd)).
-  enablePlugins(PlayScala).
-  aggregate(jsProjects.map(projectToRef): _*)
-
-lazy val js = project.enablePlugins(ScalaJSPlugin, ScalaJSPlay)
+lazy val client = project.enablePlugins(ScalaJSPlugin, ScalaJSWeb)
 ```
+Note: make sure you use the `Assets` scope.
 
-There are two auto-plugins: `ScalaJSPlay` and `PlayScalaJS`. The workflow is the following:
-* `ScalaJSPlay` plugin is enabled in Scala.js projects.
-* `PlayScalaJS` is added to the Play project. It is build on top of the SbtWeb plugin and can also add itself automatically to all projects that have SbtWeb enabled.
-* Scala.js projects are collected in the `scalaJSProjects` setting key of the Play project.
-* When compilation or testing takes place, then the `PlayScalaJS` plugin runs all required tasks on `scalaJSProjects` projects, copies the output to Play assets and takes care about source maps.
+## Examples
 
-To see the plugin in action, you can clone and run this [simple example application](https://github.com/vmunier/play-with-scalajs-example).
+To see the plugin in action, you can clone and run one of these example apps:
+- [Play with Scala.js app](https://github.com/vmunier/play-with-scalajs-example)
+- [Akka HTTP with Scala.js app](https://github.com/vmunier/akka-http-with-scalajs-example)
 
-## Features
+## How it works
 
-- Use the `scalaJSProjects` setting key to attach several Scala.js projects to the Play! project
-- The `scalaJSProd` pipeline task generates the optimised JavaScript when running `start`, `stage` and `dist`
-- Source Maps is _disabled in production_ by default to prevent your users from seeing the source files. But it can easily be enabled in production too by setting `(emitSourceMaps in fullOptJS) := true` in the Scala.js projects.
+There are two plugins: `WebScalaJS` and `ScalaJSWeb`.
+* `WebScalaJS` is automatically added to your SbtWeb project.
+* `ScalaJSWeb` should be manually added to the Scala.js projects that you want to connect to your SbtWeb project.
+* Scala.js projects are collected in the `scalaJSProjects` setting key of the SbtWeb project. The plugin does nothing if `scalaJSProjects` is not specified or is empty.
+* When compilation or testing takes place, then the `WebScalaJS` plugin runs all required tasks on `scalaJSProjects` projects, copies the output to SbtWeb assets and takes care about source maps.
+
+## Settings and Tasks
+
+* `scalaJSProjects` setting lists the Scala.js projects whose output is used by the server.
+
+* `scalaJSDev` task runs all tasks for development, including Scala.js `fastOptJS` task and source maps.
+
+* `scalaJSProd` task runs all tasks for production, including Scala.js `fullOptJS` task and source maps.
+
+* `scalaJSPipeline` task runs `scalaJSDev` when `isDevMode` is true, runs `scalaJSProd` otherwise.
+
+* `isDevMode` task returns true if the sbt command run by the user exists in the `devCommands` setting.
+  Some users may want to override `isDevMode` to read the dev/prod mode from a configuration file or from an environment variable.
+
+* `devCommands` setting contains the name of the commands used during development, which are `run`, `compile` and `re-start`.
+  It can be extended/overridden to contain different dev commands.
+
+## Source Maps
+
+The plugin copies the Scala files to the SbtWeb assets, so that they can be served to the browser and used for Source Maps.
+
+Source Map and Scala files _do not exist in production_ by default to prevent your users from seeing the source files.
+But it can easily be enabled in production too by setting `(emitSourceMaps in fullOptJS) := true` in the Scala.js projects.
+
+## Scala.js continuous compilation
+
+The plugin also watches files from the Scala.js projects.
+Redefine `compile` to trigger `scalaJSPipeline` when using `compile`, `~compile`, `~run`:
+```
+compile in Compile <<= (compile in Compile) dependsOn scalaJSPipeline.map(f => f(Seq.empty))
+```
